move add_edge_lengths to distance module
