+++ osmnx/distance.py
@@ -9,7 +9,9 @@ import pandas as pd
 from rtree.index import Index as RTreeIndex
 from shapely.geometry import Point
 
+from . import distance
 from . import projection
+from . import utils
 from . import utils_geo
 from . import utils_graph
 
@@ -98,6 +100,50 @@ def euclidean_dist_vec(y1, x1, y2, x2):
     """
     # pythagorean theorem
     return ((x1 - x2) ** 2 + (y1 - y2) ** 2) ** 0.5
+
+
+def add_edge_lengths(G, precision=3):
+    """
+    Add `length` attribute (in meters) to each edge.
+
+    Calculated via great-circle distance between each edge's incident nodes,
+    so ensure graph is in unprojected coordinates. Graph should be
+    unsimplified to get accurate distances. Note: this function is run by all
+    the `graph.graph_from_x` functions automatically to add `length`
+    attributes to all edges.
+
+    Parameters
+    ----------
+    G : networkx.MultiDiGraph
+        input graph
+    precision : int
+        decimal precision to round lengths
+
+    Returns
+    -------
+    G : networkx.MultiDiGraph
+        graph with edge length attributes
+    """
+    # extract the edges' endpoint nodes' coordinates
+    try:
+        coords = (
+            (u, v, k, G.nodes[u]["y"], G.nodes[u]["x"], G.nodes[v]["y"], G.nodes[v]["x"])
+            for u, v, k in G.edges
+        )
+    except KeyError:  # pragma: no cover
+        raise KeyError("some edges missing nodes, possibly due to input data clipping issue")
+
+    # turn the coordinates into a DataFrame indexed by u, v, k
+    cols = ["u", "v", "k", "u_y", "u_x", "v_y", "v_x"]
+    df = pd.DataFrame(coords, columns=cols).set_index(["u", "v", "k"])
+
+    # calculate great circle distances, fill nulls with zeros, then round
+    dists = distance.great_circle_vec(df["u_y"], df["u_x"], df["v_y"], df["v_x"])
+    dists = dists.fillna(value=0).round(precision)
+    nx.set_edge_attributes(G, name="length", values=dists)
+
+    utils.log("Added edge lengths to graph")
+    return G
 
 
 def nearest_nodes(G, X, Y, return_dist=False):
