+++ osmnx/graph.py
@@ -508,7 +508,7 @@ def _create_graph(response_jsons, retain_all=False, bidirectional=False):
 
     Adds length attributes in meters (great-circle distance between endpoints)
     to all of the graph's (pre-simplified, straight-line) edges via the
-    `utils_graph.add_edge_lengths` function.
+    `distance.add_edge_lengths` function.
 
     Parameters
     ----------
@@ -561,7 +561,7 @@ def _create_graph(response_jsons, retain_all=False, bidirectional=False):
 
     # add length (great-circle distance between nodes) attribute to each edge
     if len(G.edges) > 0:
-        G = utils_graph.add_edge_lengths(G)
+        G = distance.add_edge_lengths(G)
 
     return G
